:zap: refactoring into separate diffusion app
