+++ supercat/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+class DownsampleScale(Enum):
+    X2 = "X2"
+    X4 = "X4"
+
+
+class DownsampleMethod(Enum):
+    DEFAULT = "default"
+    UNKNOWN = "unknown"
+
+
